move support config to Support context
