+++ test/zoonk/support/support_notifier_test.exs
@@ -1,7 +1,7 @@
 defmodule Zoonk.Support.SupportNotifierTest do
   use Zoonk.DataCase, async: true
 
-  alias Zoonk.Config.SupportConfig
+  alias Zoonk.Support
   alias Zoonk.Support.SupportNotifier
 
   describe "deliver_support_request/2" do
@@ -12,8 +12,8 @@ defmodule Zoonk.Support.SupportNotifierTest do
       assert {:ok, email} = SupportNotifier.deliver_support_request(user_email, message)
 
       # Check email structure
-      assert email.to == [{"", SupportConfig.support_email()}]
-      assert email.from == {"Zoonk", SupportConfig.support_email()}
+      assert email.to == [{"", Support.support_email()}]
+      assert email.from == {"Zoonk", Support.support_email()}
       assert email.subject =~ "Support request"
       assert email.text_body =~ "From: #{user_email}"
       assert email.text_body =~ "Message:\n#{message}"
